Refactor directory structure for STQBM properties
