+++ qlbm/lattice/lattices/__init__.py
@@ -1,5 +1,5 @@
 from .base import Lattice
 from .collisionless_lattice import CollisionlessLattice
-from .spacetime.spacetime_lattice import SpaceTimeLattice
+from .spacetime_lattice import SpaceTimeLattice
 
 __all__ = ["Lattice", "CollisionlessLattice", "SpaceTimeLattice"]
